+++ tests/shipment_test.go
@@ -346,6 +346,7 @@ func (c *ClientTests) TestShipmentsGetNextPage() {
 }
 
 // Tests that we retrieve time-in-transit data for each of the Rates of a Shipment.
+// Deprecated: This method is deprecated and will be removed in a future release.
 func (c *ClientTests) TestShipmentGetShipmentEstimatedDeliveryDate() {
 	client := c.TestClient()
 	assert, require := c.Assert(), c.Require()
@@ -358,5 +359,47 @@ func (c *ClientTests) TestShipmentGetShipmentEstimatedDeliveryDate() {
 
 	for _, entry := range rates {
 		assert.NotNil(entry.EasyPostTimeInTransitData)
+	}
+}
+
+func (c *ClientTests) TestShipmentEstimateDeliveryDate() {
+	client := c.TestClient()
+	assert, require := c.Assert(), c.Require()
+
+	shipment, err := client.CreateShipment(c.fixture.BasicShipment())
+	require.NoError(err)
+
+	estimates, err := client.EstimateDeliveryDateForShipment(shipment.ID, &easypost.EstimateDeliveryDateForShipmentParams{
+		PlannedShipDate: c.fixture.PlannedShipDate(),
+	})
+	require.NoError(err)
+
+	assert.True(len(estimates) > 0)
+	for _, entry := range estimates {
+		assert.NotNil(entry.TimeInTransitDetails.EasyPostEstimatedDeliveryDate)
+		assert.NotNil(entry.TimeInTransitDetails.TimeInTransitPercentiles)
+		assert.NotNil(entry.TimeInTransitDetails.PlannedShipDate)
+	}
+}
+
+func (c *ClientTests) TestShipmentRecommendShipDate() {
+	client := c.TestClient()
+	assert, require := c.Assert(), c.Require()
+
+	shipment, err := client.CreateShipment(c.fixture.BasicShipment())
+	require.NoError(err)
+
+	recommendations, err := client.RecommendShipDateForShipment(shipment.ID, &easypost.RecommendShipDateForShipmentParams{
+		DesiredDeliveryDate: c.fixture.DesiredDeliveryDate(),
+	})
+	require.NoError(err)
+
+	assert.True(len(recommendations) > 0)
+	for _, entry := range recommendations {
+		assert.NotNil(entry.TimeInTransitDetails.EasyPostRecommendedShipDate)
+		assert.NotNil(entry.TimeInTransitDetails.DeliveryDateConfidence)
+		assert.NotNil(entry.TimeInTransitDetails.EstimatedTransitDays)
+		assert.NotNil(entry.TimeInTransitDetails.TimeInTransitPercentiles)
+		assert.NotNil(entry.TimeInTransitDetails.DesiredDeliveryDate)
 	}
 }